fix bug of edit universitas logo by id handler
set to only execute the delete file function on error when file in the request is exist

+++ src/Interface/http/api/universitas/handler.ts
@@ -96,7 +96,9 @@ export class UniversitasHandler {
       await this.editLogoUniversitasByIdUsecase.execute(payload);
       res.status(200).json({status: 'success'});
     } catch (e) {
-      await deleteFile(req.file?.path as string);
+      if (req.file) {
+        await deleteFile(req.file?.path as string);
+      }
       next(e);
     }
   }
